pyproject.toml
[project]
name = "langchain-rag-demo"
version = "0.1.0"
description = "A RAG (Retrieval Augmented Generation) demo using LangChain, learn what is RAG and how to use it"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-text-splitters>=0.3.0",
    "langchain-huggingface>=0.1.0",
    "langchain-chroma>=0.1.0",
    "langgraph>=0.2.0",
    "python-dotenv>=1.0.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "pytest>=8.0.0",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.4.0",
    "pytest>=8.0.0",
]
